pyproject.toml
[project]
name = "video-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "awscli-local[ver1]>=0.22.2",
    "black>=26.3.1",
    "boto3>=1.21.2",
    "botocore>=1.24.2",
    "celery[librabbitmq]>=5.6.2",
    "certifi>=2021.10.8",
    "charset-normalizer>=2.0.12",
    "idna>=3.3",
    "jmespath>=0.10.0",
    "numpy>=1.22.2",
    "opencv-contrib-python-headless>=4.9.0.80",
    "pika>=1.3.2",
    "python-dateutil>=2.8.2",
    "python-dotenv>=0.19.2",
    "requests>=2.27.1",
    "s3transfer>=0.5.1",
    "six>=1.16.0",
    "urllib3>=1.26.8",
]
